requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cors-headers==3.6.0
django-filter==2.4.0
django-phone-field==1.8.1
djangorestframework==3.12.2
drf-renderer-xlsx==0.3.9
et-xmlfile==1.0.1
importlib-metadata==3.3.0
jdcal==1.4.1
Markdown==3.3.3
openpyxl==3.0.4
pytz==2020.5
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
whitenoise==5.2.0
gunicorn==20.1.0
